requirements.txt
# مكتبات أساسية
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
python-decouple>=3.8
redis>=4.5.0
email-validator>=2.0.0

# مكتبات الذكاء الاصطناعي
openai>=0.27.0
langchain>=0.0.200
transformers>=4.28.0
mistralai>=0.1.2

# مكتبات معالجة اللغة الطبيعية
nltk>=3.8.0
spacy>=3.5.0
Qalsadi

# مكتبات WebRTC
aiortc>=1.4.0


# مكتبات واجهة المستخدم
jinja2>=3.1.0
python-multipart>=0.0.5

# مكتبات الاختبار
pytest>=7.3.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# مكتبات أخرى
python-dotenv>=1.0.0
Pillow>=9.5.0
python-dateutil>=2.8.2